move scrape logic out
